requirements.txt
# requirements.txt
# Python 3.10+ recommended

# Core
python-dotenv==1.0.0
pydantic==1.10.13

#pydantic==2.5.0
pyyaml==6.0.1

# Document Processing
pdfplumber==0.10.3
pypdf==3.17.1

# Embeddings
sentence-transformers==2.7.0
torch

# Vector Database
chromadb==0.4.10

# LLM Providers
google-generativeai==0.3.0
groq==0.4.1
ollama==0.1.0
openai==1.3.5

# Utilities
numpy==1.24.3
pandas==2.1.1
tqdm==4.66.1
requests==2.31.0

# Evaluation
scikit-learn==1.3.2

# Logging
python-json-logger==2.0.7

# Optional: Web UI
fastapi==0.99.1

#fastapi==0.104.1
uvicorn==0.24.0